backend/requirements.txt
# ============================================================================
# KetzAgenticEcomm - Backend Dependencies
# ============================================================================

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==12.0

# Azure SDKs
azure-identity==1.15.0
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
azure-ai-vision-imageanalysis==1.0.0
azure-communication-callautomation==1.1.0
openai==1.12.0

# Database
pymongo==4.6.1
motor==3.3.2

# Data Processing
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.26.0
aiohttp==3.9.3
aiofiles==23.2.1

# Image Processing
pillow==10.2.0
numpy==1.26.4

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.2.3
structlog==24.1.0

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
azure-monitor-opentelemetry==1.2.0
